requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic[email]
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
pydantic-settings==2.1.0
email-validator==2.3.0
pytest==7.4.3
httpx==0.25.2
